Adicionado classe para Combinação e uma forma de resolver fatoriais com n > 20.

+++ FuncoesUtil/Fatorial.java
@@ -1,11 +1,11 @@
 package FuncoesUtil;
 
 /**
- * Classe útil para resolver fatorial iterativamente.
+ * Classe utilitária para resolver fatorial iterativamente.
  */
 
 public class Fatorial {
-    public static int fatorial(int x) {
+    public static long fatorial(int x) {
         int fat = 1;
         for (int i = 1; i <= x; i++) {
             fat *= i;
